move modifications from viewer.py to live.py. viewer.py is now the mostly unmodified nerfview@0.0.3 viewer. see \#5
